+++ src/Pegatron/Grammars/Peg/Atom.cs
@@ -1,0 +1,24 @@
+using Pegatron.Grammars.Peg.Ast;
+using System;
+
+namespace Pegatron.Grammars.Peg
+{
+	public class Atom : IGrammarRule<INode>
+	{
+		public static string DefinitionText => "atom  :=  ruleRef | terminal | '(' choice #! ')'";
+
+		public void Register(IGrammarBuilder<INode> grammar)
+		{
+			grammar.Choice("atom",
+				grammar.Ref("ruleRef"),
+				grammar.Ref("terminal"),
+				grammar.Sequence(null, grammar.TerminalValue("("), grammar.Ref("choice").Lift(), grammar.TerminalValue(")"))
+			);
+		}
+
+		public INode Reduce(IRule rule, INodeContext<INode> page)
+		{
+			throw new NotImplementedException();
+		}
+	}
+}
